API

Purses.cache — Method

cache(x[, idx])

If x is a Purse, return its cache field, otherwise return an empty tuple. If the optional argument idx is supplied, return the cache item stored at index idx.

Examples

julia> purse = Purse(1.0, -);

julia> Purses.cache(purse, 1)
-1.0

Purses.register! — Method

register!(fs...)

Register each f in fs as cacheable and return fs.

Note

If f is already registered, register will not redefine it. It is therefore safe to call register with the same functions multiple times, without the risk of method redefinitions.

Examples

julia> purse = Purse(rand(100), Purses.register!(length)...);

julia> length(purse)
100

Purses.value — Method

value(x)

If x is a Purse, return its value field, otherwise return x itself.

Examples

julia> Purses.value(Purse(2.0))
2.0

Purses.AbstractPurse — Type

AbstractPurse{T,F<:Tuple}

Supertype for purses (or purse-like types) wrapping a value of type T and caching functions of types F. To create a subtype which conforms to the AbstractPurse interface, value and cache must be implemented.

Purses.Purse — Type

Purse{T,F<:Tuple,S<:Tuple}

Purse with a value and cache.

Fields

value::T: the value stored in the purse

cache::S: the cached values stored as a tuple in the same order as F

Purses.Purse — Method

Purse(value, fs...)

Return value wrapped in a Purse. Each f in fs can be either a callable or a pair a callable and a value. If f is a callable, the cache will automatically be calculated by applying each f to value. If f is a pair, the cache for the first element of the pair is set to the value of the second element.

Examples

julia> Purse(1.0)
Purse{Float64,Tuple{},Tuple{}}(1.0, ())

julia> Purse(2, inv)
Purse{Int64,Tuple{typeof(inv)},Tuple{Float64}}(2, (0.5,))

julia> Purse(2, inv => 0.2)
Purse{Int64,Tuple{typeof(inv)},Tuple{Float64}}(2, (0.2,))